ref(google-genai)!: Move Google GenAI integration to `@sentry/server-utils`
Move the Google GenAI gen-ai instrumentation from `@sentry/core` to
`@sentry/server-utils/src/ai/google-genai`, consuming the shared `ai/core` base.
The `instrumentGoogleGenAIClient` symbol is no longer exported from `@sentry/core`;
node/cloudflare/vercel-edge re-export it from `@sentry/server-utils`, so the platform
SDK public API is unchanged.

As with the openai/anthropic moves, `vercel-edge` re-exports from
`@sentry/server-utils/no-diagnostic-channels` to keep the edge build free of
`node:diagnostics_channel`.
