+++ packages/server-utils/src/ai/google-genai/types.ts
