+++ packages/server-utils/src/ai/google-genai/utils.ts
