+++ packages/server-utils/src/ai/index.ts
@@ -6,3 +6,4 @@
 
 export { instrumentOpenAiClient } from './openai';
 export { instrumentAnthropicAiClient } from './anthropic-ai';
+export { instrumentGoogleGenAIClient } from './google-genai';
