+++ packages/server-utils/src/integrations/tracing-channel/google-genai.ts
@@ -1,20 +1,18 @@
 import { GEN_AI_REQUEST_MODEL } from '@sentry/conventions/attributes';
 import * as diagnosticsChannel from 'node:diagnostics_channel';
-import type { GoogleGenAIOptions, GoogleGenAIResponse, IntegrationFn, Span } from '@sentry/core';
+import type { IntegrationFn, Span } from '@sentry/core';
 import {
   _INTERNAL_shouldSkipAiProviderWrapping,
-  addGoogleGenAIRequestAttributes,
-  addGoogleGenAIResponseAttributes,
   defineIntegration,
-  extractGoogleGenAIRequestAttributes,
   getActiveSpan,
-  instrumentGoogleGenAIStream,
-  resolveAIRecordingOptions,
   SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN,
-  shouldEnableTruncation,
   spanToJSON,
   startInactiveSpan,
 } from '@sentry/core';
+import { resolveAIRecordingOptions, shouldEnableTruncation } from '../../ai/core/utils';
+import { addPrivateRequestAttributes, addResponseAttributes, extractRequestAttributes } from '../../ai/google-genai';
+import { instrumentStream } from '../../ai/google-genai/streaming';
+import type { GoogleGenAIOptions, GoogleGenAIResponse } from '../../ai/google-genai/types';
 import { CHANNELS } from '../../orchestrion/channels';
 import { bindTracingChannelToSpan } from '../../tracing-channel';
 import { googleGenAiModuleNames } from '../../orchestrion/config/google-genai';
@@ -59,7 +57,7 @@ function instrumentGoogleGenai(options: GoogleGenAIOptions): void {
         beforeSpanEnd: (span, data) => {
           // Embeddings responses carry no content attributes.
           if (operation !== 'embeddings') {
-            addGoogleGenAIResponseAttributes(
+            addResponseAttributes(
               span,
               data.result as GoogleGenAIResponse,
               resolveAIRecordingOptions(options).recordOutputs,
@@ -106,7 +104,7 @@ function createGenAiSpan(
   const { recordInputs } = resolveAIRecordingOptions(options);
   const enableTruncation = shouldEnableTruncation(options.enableTruncation);
 
-  const attributes = extractGoogleGenAIRequestAttributes(operation, params, data.self);
+  const attributes = extractRequestAttributes(operation, params, data.self);
   const model = (attributes[GEN_AI_REQUEST_MODEL] as string) || 'unknown';
   attributes[SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN] = ORIGIN;
 
@@ -117,7 +115,7 @@ function createGenAiSpan(
   });
 
   if (recordInputs && params) {
-    addGoogleGenAIRequestAttributes(span, params, operation, enableTruncation);
+    addPrivateRequestAttributes(span, params, operation, enableTruncation);
   }
 
   return span;
@@ -131,7 +129,7 @@ function isAsyncIterable(value: unknown): value is AsyncIterableStream {
 
 /**
  * Only the streaming methods (`generateContentStream`/`sendMessageStream`) resolve to an async iterable.
- * For a stream we patch `result[Symbol.asyncIterator]` in place so `instrumentGoogleGenAIStream` ends the
+ * For a stream we patch `result[Symbol.asyncIterator]` in place so `instrumentStream` ends the
  * span when iteration finishes.
  */
 function wrapStreamResult(span: Span, data: GoogleGenAIChannelContext, options: GoogleGenAIOptions): boolean {
@@ -142,7 +140,7 @@ function wrapStreamResult(span: Span, data: GoogleGenAIChannelContext, options: 
 
   const { recordOutputs } = resolveAIRecordingOptions(options);
   const iterate = result[Symbol.asyncIterator].bind(result);
-  const instrumented = instrumentGoogleGenAIStream({ [Symbol.asyncIterator]: iterate }, span, recordOutputs ?? false);
+  const instrumented = instrumentStream({ [Symbol.asyncIterator]: iterate }, span, recordOutputs ?? false);
   result[Symbol.asyncIterator] = () => instrumented;
 
   return true;
